+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Development
 
+## Version 3.2.1
+
+- Added `child_label` to entry serializers metadata.
+
+## Version 3.2.0
+
+- Added `filter_weighted` flag to `TunicClient.get_campaigns()`
+
 ## Version 3.1.3
 
 - Added `AbstractInfographic.get_infographic_type_name` method.